requirements.txt
Django>=4.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
django-environ>=0.10.0
py-bcrypt>=0.4
